Okay, borked something trying to get named-parameters+closure but borked something along the way.  Dunno what.
Trying to pass, now:

  TupleExpression( MapExpression, Closure )

for your wacky closure sugar, but it ain't wanting to play friendly.

So, I think non-named work, named seem to maybe work, but the closure is
getting lost?  AntBuilder is getting a request for a 'doCall' datatype
which is clearly wrong.

Anyhow, b0rked.  Offline soon.  Pass the hot potato.