Changes to the code base as per JIRA issue: http://jira.codehaus.org/browse/GROOVY-1690
We add an adapter factory to IJavaProjects that will create adapters that adapt from the IJavaProject to either 
a GroovyProject or an IActionFilter. 

The IActionFilter checks to see if the IProject has the Groovy Nature defined and if so, returns true.

An action ImportLibs was added that is enabled when the adapted IActionFilter returns true, or when the
IProject is both an IJavaProject and a GroovyProject.  The ImportLibs action then copies the jar libraries 
from the plugin installation to the workspace project, directly importing them.

GroovyRuntime was modified to include methods to not just add the jars as links to a project, but to copy them over
in support of the action ImportLibs.

I still need to write a couple of unit tests (BAD BAD ME) to simply demonstrate that this function works correctly, 
but I did test it out a number of times.